1. Fix error in transport start/stop logic. Transport was started before connection to    the module was established, therefore many first buffers were lost. Was happened, while    transport was reacting on connection activating with memory pool. 2. Fix flushing error in hadaq::MbsTransmitterModule. When EOF marker was appearing,    rest of data in output iterator was not finalized and was not delivered to the output.
git-svn-id: https://subversion.gsi.de/dabc/trunk@1577 bcbf6573-9a26-0410-9ebc-ce4ab7aade96

+++ plugins/hadaq/src/HldFile.cxx
@@ -84,7 +84,7 @@ bool hadaq::HldFile::OpenRead(const char* fname)
 //   DOUT0("Open HLD file %s for reading", fname);
 
    hadaq::Event evnt;
-   uint64_t size = sizeof(hadaq::Event);
+   uint32_t size = sizeof(hadaq::Event);
 
    if (!ReadBuffer(&evnt, &size, true)) {
       EOUT("Cannot read starting event from file");
@@ -137,7 +137,7 @@ bool hadaq::HldFile::WriteBuffer(void* buf, uint32_t bufsize)
    return true;
 }
 
-bool hadaq::HldFile::ReadBuffer(void* ptr, uint64_t* sz, bool onlyevent)
+bool hadaq::HldFile::ReadBuffer(void* ptr, uint32_t* sz, bool onlyevent)
 {
    if (!isReading() || (ptr==0) || (sz==0) || (*sz < sizeof(hadaq::HadTu))) return false;
